feat(Show Group Work): Show component work in Grading Tool
#1037
