+++ src/assets/wise5/components/showGroupWork/show-group-work-grading/show-group-work-grading.component.html
@@ -1,4 +1,3 @@
-<component-header [component]="component"></component-header>
 <show-group-work-display
   [componentId]="componentId"
   [componentContent]="componentContent"
