+++ src/assets/wise5/components/showGroupWork/show-group-work-grading/show-group-work-grading.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { ShowGroupWorkDisplayModule } from '../show-group-work-display/show-group-work-display.module';
+import { ShowGroupWorkGradingComponent } from './show-group-work-grading.component';
+
+@NgModule({
+  declarations: [ShowGroupWorkGradingComponent],
+  imports: [ShowGroupWorkDisplayModule],
+  exports: [ShowGroupWorkGradingComponent]
+})
+export class ShowGroupWorkGradingModule {}
